+++ src/artgraine/model/PersistableInterface.java
@@ -6,5 +6,5 @@ public interface PersistableInterface<T> {
 
     public int save(T entity)throws SQLException;
 
-    public int deleteOneById(int id) throws SQLException;
+    public int deleteOneById(Long id) throws SQLException;
 }
